update: refactoring based on new folder structure
