+++ flask-restfull/security.py
@@ -1,5 +1,5 @@
 from werkzeug.security import safe_str_cmp
-from user import User
+from resources.user import User
 
 
 def authenticate(username, password):
